pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-ai"
version = "0.1.0"
description = "AI-powered git commit message generator"
authors = [{name = "Aju John", email = "aju@ajujohn.me"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.1.0",
    "gitpython>=3.1.40",
    "anthropic>=0.21.0",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
git-ai = "git_ai.cli:main"

[project.urls]
Homepage = "https://github.com/ajujohn/git-ai"
Repository = "https://github.com/ajujohn/git-ai"
